Arc Web 20-10

+++ ArquiteturaWEB/ac2/src/main/java/com/example/ac2/repository/CursoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ac2.models.Curso;
 
-public interface CursoRepository extends JpaRepository<Curso, Long> {
+public interface CursoRepository extends JpaRepository<Curso, Integer> {
 
 }
